Fixed iterators/order. Added backward iterators.
git-svn-id: http://hdt-it.googlecode.com/svn/trunk@162 ca5865f0-ea67-0c57-7e18-94a3701c88e0

+++ hdt-lib/src/dictionary/PFCDictionary.hpp
@@ -50,8 +50,8 @@ private:
 	csd::CSD *shared;
 
 	unsigned int mapping;
-	size_t sizeStrings;
-	size_t blocksize;
+	uint64_t sizeStrings;
+	uint32_t blocksize;
 
 	//ControlInformation controlInformation;
 	HDTSpecification spec;
